requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cmake==3.31.4
colorama==0.4.6
cryptography==44.0.0
diskcache==5.6.3
et_xmlfile==2.0.0
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.28.1
idna==3.10
inquirerpy==0.3.4
Jinja2
joblib==1.4.2
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.3
numpy==1.26.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.5
pfzy==0.3.4
pillow==10.2.0
prompt_toolkit==3.0.50
pybind11==2.13.6
pycparser==2.22
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scikit_build_core==0.10.7
scipy==1.15.1
sentence-transformers==3.4.1
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
xlrd==2.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cmake==3.31.4
colorama==0.4.6
cryptography==44.0.0
diskcache==5.6.3
et_xmlfile==2.0.0
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.28.1
idna==3.10
inquirerpy==0.3.4
Jinja2==3.1.3
joblib==1.4.2
llama_cpp_python @ https://github.com/abetlen/llama-cpp-python/releases/download/v0.3.4-cu124/llama_cpp_python-0.3.4-cp312-cp312-linux_x86_64.whl
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.3
numpy==1.26.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.5
pfzy==0.3.4
pillow==10.2.0
prompt_toolkit==3.0.50
pybind11==2.13.6
pycparser==2.22
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scikit_build_core==0.10.7
scipy==1.15.1
sentence-transformers==3.4.1
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
xlrd==2.0.1
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
colorama==0.4.6
cryptography==44.0.0
diskcache==5.6.3
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
inquirerpy==0.3.4
Jinja2==3.1.5
joblib==1.4.2
llama_cpp_python @ https://github.com/abetlen/llama-cpp-python/releases/download/v0.3.4-cu124/llama_cpp_python-0.3.4-cp312-cp312-linux_x86_64.whl#sha256=c3e1cab50aca2194d55c4164c2d01304919f0533e9a9aedb1390c2bfd95dd6d0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.3
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pfzy==0.3.4
pillow==10.2.0
prompt_toolkit==3.0.50
pybind11==2.13.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
safetensors==0.5.2
scikit-learn==1.6.1
scikit_build_core==0.10.7
scipy==1.15.1
sentence-transformers==3.4.1
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
triton==3.2.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
wheel==0.45.1
xlrd==2.0.1
